requirement.txt
boto3==1.20.24
requests==2.31.0
line-bot-sdk==3.5.0
python-dotenv==0.19.1
